add CollectionDefinition and refactor stage 1
